Commented out desired but not yet implemented APIs
